requirements.txt
Flask==3.1.0
Flask_MySQLdb==2.0.0
GitPython==3.1.43
GitPython==3.1.44
ollama==0.4.7
pydantic==2.10.6
Requests==2.32.3
Werkzeug==3.1.3

Be@awsuser846965